requirements.txt
git+https://github.com/ninjiez/discord.py-self@renamed#egg=selfcord.py-self
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
decorator==5.1.1
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
nest-asyncio==1.5.5
requests==2.27.1
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
validators==0.19.0
yarl==1.7.2
